+++ src/main/kotlin/com/github/liaoxiangyun/ideaplugin/coderaminder/model/GitSummary.kt
@@ -1,5 +1,7 @@
 package com.github.liaoxiangyun.ideaplugin.coderaminder.model
 
+import com.fasterxml.jackson.annotation.JsonProperty
+import com.github.liaoxiangyun.ideaplugin.coderaminder.util.CalendarUtil
 import java.time.LocalDate
 import java.time.LocalDateTime
 
@@ -22,11 +24,13 @@ class GitSummary {
 
     class Day {
         var date: LocalDate
+        var dateStr: String
         var epochDay: Long
         var total: Int = 0
 
         constructor(date: LocalDate, epochDay: Long, total: Int) {
             this.date = date
+            this.dateStr = CalendarUtil.format(date)
             this.epochDay = epochDay
             this.total = total
         }
@@ -38,6 +42,7 @@ class GitSummary {
         var projectName: String = ""
         var createAt: LocalDateTime
         var date: LocalDate
+        var dateStr: String
         var epochDay: Long
         var add: Int = 0
         var del: Int = 0
@@ -53,6 +58,7 @@ class GitSummary {
             this.total = total
             val localDate = createAt.toLocalDate()
             this.date = localDate
+            this.dateStr = CalendarUtil.format(localDate)
             epochDay = localDate.toEpochDay()
         }
     }