+++ src/main/kotlin/com/github/liaoxiangyun/ideaplugin/coderaminder/model/gitlab/Author.kt
@@ -1,0 +1,6 @@
+package com.github.liaoxiangyun.ideaplugin.coderaminder.model.gitlab
+
+class Author {
+    var name = ""
+    var email = ""
+}