+++ src/main/kotlin/com/github/liaoxiangyun/ideaplugin/coderaminder/util/CalendarUtil.kt
@@ -1,8 +1,7 @@
 package com.github.liaoxiangyun.ideaplugin.coderaminder.util
 
+import cn.hutool.json.JSONUtil
 import com.github.liaoxiangyun.ideaplugin.coderaminder.settings.CodeSettingsState
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
 import java.io.BufferedReader
 import java.io.InputStreamReader
 import java.text.ParseException
@@ -15,15 +14,99 @@ object CalendarUtil {
     @Throws(ParseException::class)
     @JvmStatic
     fun main(args: Array<String>) {
-        val content = getContent("/json/2022.json")
-        println("content = ${content}")
 
+        val weekDays = getWeekDays(1)
+        println("weekDays = ${weekDays}")
 
     }
 
-    private val gson = Gson()
-
     open val calendarMap = mutableMapOf<String, Int>()
+
+
+    private val dateFormat: DateTimeFormatter = DateTimeFormatter.ofPattern("yyyyMMdd")
+    open fun format(date: LocalDate): String {
+        return date.format(dateFormat)
+    }
+
+    open fun isOffDay(localDate: LocalDate): Boolean {
+        val key = localDate.format(dateFormat)
+        if (calendarMap.containsKey(key)) {
+            val i = calendarMap[key]
+            if (i !== 1) {
+                return true
+            }
+        } else {
+            val week: DayOfWeek = localDate.dayOfWeek
+            if (week == DayOfWeek.SATURDAY || week == DayOfWeek.SUNDAY) {
+                return true
+            }
+        }
+        return false
+    }
+
+    open fun getWeekDays(week: Int): ArrayList<LocalDate> {
+        loadData()
+        val now = LocalDate.now()
+        var weekDays = getWeekDays2(now)
+        if (week == 0) {
+            return weekDays
+        }
+        if (week > 0) {
+            for (index in 1..week) {
+                println("#getWeekDays week=$week index=$index")
+                weekDays = getWeekDays2(weekDays.last().plusDays(1))
+            }
+        } else {
+            for (index in 1..-week) {
+                println("#getWeekDays week=$week index=$index")
+                weekDays = getWeekDays2(weekDays.first().plusDays(-1))
+            }
+        }
+        return weekDays
+    }
+
+    private fun getWeekDays2(date: LocalDate): ArrayList<LocalDate> {
+        val list = arrayListOf<LocalDate>()
+        var begin: LocalDate = date
+        var end: LocalDate = date
+        //向前查
+        var max = 10
+        while (true) {
+            val b = isOffDay(begin)
+            val last = begin.plusDays(-1)
+            val b1 = isOffDay(last)
+//            println("$begin b=$b ; last b1=$b1 ; ${if (!b && b1) "r=true" else ""}")
+            if (!b && b1) {
+                break
+            }
+            begin = last
+            if (--max <= 0) {
+                break
+            }
+        }
+        max = 10
+        //向后查
+        while (true) {
+            val b = isOffDay(end)
+            val next = end.plusDays(1)
+            val b1 = isOffDay(next)
+//            println("$end b=$b ; next b1=$b1 ; ${if (b && !b1) "r=true" else ""}")
+            if (b && !b1) {
+                break
+            }
+            end = next
+            if (--max <= 0) {
+                break
+            }
+        }
+        do {
+            list.add(begin)
+            begin = begin.plusDays(1)
+        } while (begin <= end)
+
+        return list
+    }
+
 
     private fun getContent(path: String): String {
         val stream = CalendarUtil.javaClass.getResourceAsStream(path)
@@ -43,107 +126,35 @@ object CalendarUtil {
     }
 
 
-    private fun <T> toObj(json: String): T {
-        return gson.fromJson(json, object : TypeToken<T>() {}.type)
+    private fun <T> getObj(json: String, clazz: Class<T>): T {
+        return JSONUtil.toBean(json, clazz)
     }
 
-    open fun load(): MutableMap<String, Int> {
-        val content = getContent("/json/2022.json")
-        println("content = $content")
-
-        val toObj = toObj<Map<String, Map<String, Int>>>(content)
-        for (entry in toObj) {
-            for (e in entry.value) {
-                calendarMap[e.key] = e.value
-            }
-        }
-
+    private fun setMapByJson(str: String): MutableMap<String, Int> {
         try {
-            val str = CodeSettingsState.instance.calendar
-            if (str.trim().isNotBlank()) {
-                val toObj1 = toObj<Map<String, Map<String, Int>>>(str)
-                for (entry in toObj1) {
-                    for (e in entry.value) {
-                        calendarMap[e.key] = e.value
+            if (str.isNotBlank()) {
+                val toObj = getObj(str, Map::class.java)
+                for (entry in toObj) {
+                    for (e in (entry.value as Map<Any, Any>)) {
+                        calendarMap[e.key as String] = e.value.toString().toInt()
                     }
                 }
             }
-
         } catch (e: Exception) {
         }
         return calendarMap
     }
 
+    private fun loadData() {
+        val settingsState = CodeSettingsState.instance
+        if (settingsState.calendar.isBlank()) {
+            settingsState.calendar = getContent("/json/2022.json")
+        }
+        setMapByJson(settingsState.calendar)
+    }
+
     init {
-        val load = load()
-        println(load)
-    }
-
-
-    private val dateFormat: DateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
-    private fun isOffDay(localDate: LocalDate): Boolean {
-        val key = localDate.format(dateFormat)
-        if (calendarMap.containsKey(key)) {
-            val i = calendarMap[key]
-            if (i !== 1) {
-                return true
-            }
-        } else {
-            val week: DayOfWeek = localDate.dayOfWeek
-            if (week == DayOfWeek.SATURDAY || week == DayOfWeek.SUNDAY) {
-                return true
-            }
-        }
-        return false
-    }
-
-    open fun getWeekDays(week: Int): ArrayList<LocalDate> {
-        val now = LocalDate.now()
-        var weekDays = getWeekDays(now)
-
-        for (index in week..0) {
-            weekDays = if (index > 0) {
-                getWeekDays(weekDays.last().plusDays(1))
-            } else if (index < 0) {
-                getWeekDays(weekDays.first().plusDays(-1))
-            } else {
-                break;
-            }
-        }
-
-        return weekDays
-    }
-
-    open fun getWeekDays(date: LocalDate): ArrayList<LocalDate> {
-        val list = arrayListOf<LocalDate>()
-        var begin: LocalDate = date
-        var end: LocalDate = date
-        //向前查
-        while (true) {
-            val b = isOffDay(begin)
-            val last = begin.plusDays(-1)
-            val b1 = isOffDay(last)
-            if (!b && b1) {
-                break
-            }
-            begin = last
-        }
-        //向后查
-        while (true) {
-            val b = isOffDay(end)
-            val next = begin.plusDays(1)
-            val b1 = isOffDay(next)
-            if (b && !b1) {
-                break
-            }
-            end = next
-        }
-        do {
-            list.add(begin)
-            begin.plusDays(1)
-        } while (begin < end)
-
-        return list
+        loadData()
     }
 
 }