+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [Unreleased]
 ### Added
-- 2022-04-06 : 代码量提醒 [liaoxy-helper-v1.2.0](https://github.com/liaoxiangyun/liaoxy-idea-helper/releases/tag/1.x)
+- 2022-04-06 : 代码量统计 [liaoxy-helper-v1.2.0](https://github.com/liaoxiangyun/liaoxy-idea-helper/releases/tag/1.x)
 - 2022-02-14 : java自动补全
 - 2021-12-30 : 修改一些Bug
 - 2021-08-12 : umi dispatch 导航